Added slides coroutine to tutorial manager. Another coroutine scans through the list of slide objects which contain a filename, timer (how long it displays), and text. It will call the slides coroutine on all objects, drawing the image for a specified amount of time along with a tutorial box with the text.

+++ KarmaTrauma/New Unity Project 1/Assets/Scripts/Managers/GameManager.cs
@@ -246,7 +246,7 @@ public class GameManager : MonoBehaviour
 		}
     }
 
-    public GameObject CreateTutorialBox(string message, int destroyTimer = -1, Textbox.TutorialBoxPosition position = Textbox.TutorialBoxPosition.MIDDLE)
+    public GameObject CreateTutorialBox(string message, float destroyTimer = -1, Textbox.TutorialBoxPosition position = Textbox.TutorialBoxPosition.MIDDLE)
     {
         GameObject dialog = (GameObject)Instantiate(dialogueContainer, dialogueContainer.transform.position, Quaternion.identity);
         StartCoroutine(dialog.GetComponent<Textbox>().DrawTutorialBox(message, destroyTimer, position));
